fix: GPT-2 end-to-end inference — 4 critical bugs
1. MatMul 3D batch dim: flatten batch dims into M for 3D+ inputs
   (was using shape[0] as M, losing batch*seq dimension)

2. Tanh numel: add OP_TANH to graph_execute's numel setup
   (Tanh kernel reads numel from inputs[1], was missing from dispatch)

3. Graph input dtype: parse ONNX elem_type from ValueInfoProto
   (was always creating F32 input tensors, now respects model dtype)

4. BERT test: update input tensors to I64 to match ONNX model dtype

GPT-2 test: max_diff 1.90 → 0.42, next_token 214 vs ref 212.
All 34 tests pass.

+++ src/operator/nn/gather.c
@@ -17,7 +17,7 @@ int gather_f32(const void* inputs[], void* outputs[],
     int64_t ni = p->num_indices;
     int64_t bs = p->block_size;
     int64_t os = p->outer_size;
-    int indices_are_i64 = p->out_axis_dim;  /* reuse field as flag */
+    int indices_are_i64 = (int)p->out_axis_dim;  /* reuse field as flag */
 
     for (int64_t o = 0; o < os; o++) {
         for (int64_t i = 0; i < ni; i++) {
